# Conflicts:
#	.travis.yml
#	README.md
#	appveyor.yml

+++ src/game/items/CItemBase.cpp
@@ -1,17 +1,16 @@
-
 // define the base types of an item (rather than the instance)
 
+#include "../../common/CUOInstall.h"
+#include "../../common/CLog.h"
+#include "../../common/CException.h"
 #include "../../sphere/ProfileTask.h"
 #include "../components/CCPropsItem.h"
 #include "../components/CCPropsItemChar.h"
 #include "../components/CCPropsItemEquippable.h"
 #include "../components/CCPropsItemWeapon.h"
 #include "../components/CCPropsItemWeaponRanged.h"
-#include "../resource/CResourceLock.h"
+#include "../../common/resource/CResourceLock.h"
 #include "../uo_files/CUOItemInfo.h"
-#include "../CUOInstall.h"
-#include "../CLog.h"
-#include "../CException.h"
 #include "CItemBase.h"
 
 
@@ -2151,7 +2150,7 @@ bool CItemBaseMulti::r_WriteVal(lpctstr ptcKey, CSString & sVal, CTextConsole * 
                     return false;
 
                 SKIP_SEPARATORS(ptcKey);
-                const CMultiComponentItem& item = m_Components[iIndex];
+                const CMultiComponentItem& item = m_Components[(size_t)iIndex];
 
                 if (!strnicmp(ptcKey, "ID", 2)) sVal.FormatVal(item.m_id);
                 else if (!strnicmp(ptcKey, "DX", 2)) sVal.FormatVal(item.m_dx);
